fix(frontend): poll containers while event stream is disconnected
The polling fallback was inverted: with the event stream disconnected
the containers query had no refetch interval at all, reintroducing the
stale-state issue (#61) exactly when polling matters most. Poll every
30s when disconnected and keep the 60s safety net when connected.

The container detail page fetched the containers query inline without
any event subscription or polling, so it also went stale. Extract the
event-driven wiring (debounced invalidation + fallback poll) from the
dashboard into a shared useLiveContainersQuery hook and use it on both
the dashboard and the detail page.

Fixes #61
